✅ Make Sum Divisible by P | Medium | Java | Runtime: 28 ms (54.28%) | Memory: 98.7 MB | 2026-08-23

+++ medium/1590-make-sum-divisible-by-p/solution.java
@@ -14,11 +14,11 @@ class Solution {
        long sum=0;
        for(int i=0;i<nums.length;i++){
            sum+=nums[i];
-           int req=(sum%p-target+p)%p;
+           int req=(int)((sum%p-target+p)%p);
            if(map.containsKey(req)){
             min=Math.min(min,i-map.get(req));
            }
-           map.put(sum%p,i);
+           map.put((int)(sum%p),(i));
        }
        if(min==Integer.MAX_VALUE || min==nums.length){
         return -1;
